lista de pokemons estilizada
