+++ pokedex/src/Pages/Pokedex/Pokedex.js
@@ -1,0 +1,23 @@
+import React from "react";
+import { useHistory } from "react-router-dom";
+import { goToPokemonDetails, goToPokemonList } from "../../routes/Coordinator";
+
+const Pokedex = () => {
+  const history = useHistory();
+
+  return (
+    <>
+      <p> AQUI É A POKEDEX </p>
+      <button onClick={() => goToPokemonList(history)}>
+        {" "}
+        Voltar para lista dos Pokemons{" "}
+      </button>
+      <button onClick={() => goToPokemonDetails(history)}>
+        {" "}
+        Ver detalhes do Pokemon{" "}
+      </button>
+    </>
+  );
+};
+
+export default Pokedex;
