--- pokedex/src/Pages/PokedexMain/MainStyle.js
@@ -1,14 +1,0 @@
-import styled from 'styled-components'
-
-export const Container = styled.section`
-display:flex;
-justify-content:space-around;
-flex-wrap:wrap;
-
-`
-export const Card = styled.article`
-background:LightSalmon;
-width:8rem;
-height:10rem;
-margin:4px;
-`
