+++ pokedex/src/Pages/PokemonList/Styled.js
@@ -1,0 +1,18 @@
+import styled from 'styled-components';
+
+export const ListWrapper = styled.section`
+  display: flex;
+  flex-wrap: wrap;
+  justify-content: space-around;
+  width: 100vw;
+`
+
+export const Card = styled.article`
+  align-items: center;
+  background: lightSalmon;
+  display: flex;
+  flex-direction: column;
+  height: 200px;
+  margin: 8px;
+  width: 160px;
+`
